Add plot_comparison and cross_correlation procedures to .biosignals
